requirements.txt
openai-agents>=0.1.0
pinecone>=5.0.0
PyMuPDF>=1.23.0
unstructured>=0.12.0
redis>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
rich>=13.0.0
